move User to app/Models
